cgroup2: implement closing cgroupsv2.Manager to prevent inotify instance leak
When watcher.Run() is terminated through the context object, watcher.Close() is
called but this is a no-op. This leaks the EventChannel routine, and the
associated inotify instance. This can be observed with crashing containers in
k8s, where more and more inotify instances are consumed.

containerd/cgroups now exposes a channel that we can close to signal the
EventChannel routine that it should terminate. This prevents the inotify leak.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>